server/requirements.txt
Flask
Flask-Cors
Flask-Mail
Flask-JWT-Extended
python-dotenv
PyMuPDF
spacy
scikit-learn
google-generativeai